+++ Block/Product.php
@@ -10,7 +10,7 @@ use Magento\Framework\View\Element\Template;
 /**
  * Class Product
  *
- * @version 1.0.6
+ * @version 1.0.7
  * @author Improntus <http://www.improntus.com> - Ecommerce done right
  * @copyright Copyright (c) 2020 Improntus
  * @package Improntus\RetailRocket\Block