+++ Model/Session.php
@@ -7,7 +7,7 @@ use Magento\Framework\Session\SessionManager;
 /**
  * Class Session
  *
- * @version 1.0.6
+ * @version 1.0.7
  * @author Improntus <http://www.improntus.com> - Ecommerce done right
  * @copyright Copyright (c) 2020 Improntus
  * @package Improntus\RetailRocket\Model