+++ Observer/NewsletterSubscriberSave.php
@@ -14,7 +14,7 @@ use Magento\Customer\Model\Customer;
 /**
  * Class NewsletterSubscriberSave
  *
- * @version 1.0.6
+ * @version 1.0.7
  * @author Improntus <http://www.improntus.com> - Ecommerce done right
  * @copyright Copyright (c) 2020 Improntus
  * @package Improntus\RetailRocket\Observer